move config into separate file to allow for future changes
- also fixed a warning

+++ src/math/vec3.c
@@ -83,7 +83,7 @@ Vec3 Vec3Normalize(const Vec3* v)
 {
 	// naive check for zero vector
 	float magnitude = Vec3Lenth(v);
-	if (fabs(magnitude) < AML_EPSILON)
+	if (fabsf(magnitude) < AML_EPSILON)
 	{
 		return (Vec3) { NAN, NAN, NAN };
 	}
